+++ jave-core/src/main/java/ws/schild/jave/VideoFilter.java
@@ -8,6 +8,10 @@ public class VideoFilter {
 
     private String expression;
 
+    public VideoFilter() {
+        this.expression = "";
+    }
+    
     public VideoFilter(String expression) {
         this.expression = expression;
     }