+++ jave-core/src/main/java/ws/schild/jave/VideoFilter.java
@@ -20,8 +20,9 @@ public class VideoFilter {
         return this.expression;
     }
 
-    public void setExpression(String expression) {
+    public VideoFilter setExpression(String expression) {
         this.expression = expression;
+        return this;
     }
 
     @Override